+++ tests/test_updates.py
@@ -1,0 +1,97 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+""" Unit tests for zeroconf._services. """
+
+import logging
+import socket
+import time
+import os
+import unittest
+from threading import Event
+
+import pytest
+
+import zeroconf as r
+from zeroconf import const
+from zeroconf import Zeroconf
+from zeroconf._services.browser import ServiceBrowser
+from zeroconf._services.info import ServiceInfo
+
+from . import has_working_ipv6, _clear_cache
+
+
+log = logging.getLogger('zeroconf')
+original_logging_level = logging.NOTSET
+
+
+def setup_module():
+    global original_logging_level
+    original_logging_level = log.level
+    log.setLevel(logging.DEBUG)
+
+
+def teardown_module():
+    if original_logging_level != logging.NOTSET:
+        log.setLevel(original_logging_level)
+
+
+def test_legacy_record_update_listener():
+    """Test a RecordUpdateListener that does not implement update_records."""
+
+    # instantiate a zeroconf instance
+    zc = Zeroconf(interfaces=['127.0.0.1'])
+
+    with pytest.raises(RuntimeError):
+        r.RecordUpdateListener().update_record(
+            zc, 0, r.DNSRecord('irrelevant', const._TYPE_SRV, const._CLASS_IN, const._DNS_HOST_TTL)
+        )
+
+    updates = []
+
+    class LegacyRecordUpdateListener(r.RecordUpdateListener):
+        """A RecordUpdateListener that does not implement update_records."""
+
+        def update_record(self, zc: 'Zeroconf', now: float, record: r.DNSRecord) -> None:
+            nonlocal updates
+            updates.append(record)
+
+    listener = LegacyRecordUpdateListener()
+
+    zc.add_listener(listener, None)
+
+    # dummy service callback
+    def on_service_state_change(zeroconf, service_type, state_change, name):
+        pass
+
+    # start a browser
+    type_ = "_homeassistant._tcp.local."
+    name = "MyTestHome"
+    browser = ServiceBrowser(zc, type_, [on_service_state_change])
+
+    info_service = ServiceInfo(
+        type_,
+        '%s.%s' % (name, type_),
+        80,
+        0,
+        0,
+        {'path': '/~paulsm/'},
+        "ash-2.local.",
+        addresses=[socket.inet_aton("10.0.1.2")],
+    )
+
+    zc.register_service(info_service)
+
+    zc.wait(1)
+
+    browser.cancel()
+
+    assert len(updates)
+    assert len([isinstance(update, r.DNSPointer) and update.name == type_ for update in updates]) >= 1
+
+    zc.remove_listener(listener)
+    # Removing a second time should not throw
+    zc.remove_listener(listener)
+
+    zc.close()
