+++ zeroconf/__init__.py
@@ -49,7 +49,6 @@ from ._protocol import DNSIncoming, DNSOutgoing  # noqa # import needed for back
 from ._services import (  # noqa # import needed for backwards compat
     Signal,
     SignalRegistrationInterface,
-    RecordUpdateListener,
     ServiceListener,
     ServiceStateChange,
 )
@@ -62,6 +61,7 @@ from ._services.info import (  # noqa # import needed for backwards compat
 )
 from ._services.registry import ServiceRegistry  # noqa # import needed for backwards compat
 from ._services.types import ZeroconfServiceTypes
+from ._updates import RecordUpdateListener, RecordUpdate
 from ._utils.name import service_type_name  # noqa # import needed for backwards compat
 from ._utils.net import (  # noqa # import needed for backwards compat
     add_multicast_member,
