+++ zeroconf/_core.py
@@ -34,14 +34,15 @@ from typing import Dict, List, Optional, Tuple, Type, Union, cast
 from ._cache import DNSCache
 from ._dns import DNSQuestion, DNSQuestionType
 from ._exceptions import NonUniqueNameException
-from ._handlers import QueryHandler, RecordManager, RecordUpdate
+from ._handlers import QueryHandler, RecordManager
 from ._history import QuestionHistory
 from ._logger import QuietLogger, log
 from ._protocol import DNSIncoming, DNSOutgoing
-from ._services import RecordUpdateListener, ServiceListener
+from ._services import ServiceListener
 from ._services.browser import ServiceBrowser
 from ._services.info import ServiceInfo, instance_name_from_service_info
 from ._services.registry import ServiceRegistry
+from ._updates import RecordUpdate, RecordUpdateListener
 from ._utils.aio import get_running_loop, shutdown_loop, wait_event_or_timeout
 from ._utils.name import service_type_name
 from ._utils.net import (
